refactored card relationship and implement switching cards in wallet

+++ Corzbank.Services/UserService.cs
@@ -60,6 +60,7 @@ namespace Corzbank.Services
                 return null;
 
             var tokenForDeleting = await _genericService.FindByCondition(x => x.User.Id == user.Id);
+            if(tokenForDeleting != null)
             await _genericService.Remove(tokenForDeleting);
 
             var tokens = new Token
